requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
Django==3.1.3
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-filter==2.4.0
djangorestframework==3.12.2
drf-schema-adapter==2.0.7
Inflector==3.0.1
ipykernel==4.6.1
mysqlclient==2.0.1
notebook==5.2.2
Pillow==8.0.1
pytz==2020.4
soupsieve==2.0.1
sqlparse==0.4.1
wincertstore==0.2
